bugfix: added term to shear estimator with l1 <--> l2 swapped
The auto- and cross-spectra (with the true lensing potential) of shear reconstructions were previously biased low. With this bugfix, we recover the expected behaviour. It might not be the fix that best fits the style of the code, but I hope it is a good starting point for a more elegant and permanent one.